+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 02.06.2026, Version 2.21.0
+
+- feature/alert-action-alert-source-relation in [#136](https://github.com/iLert/terraform-provider-ilert/pull/136)
+
 ## 22.04.2026, Version 2.20.0
 
 - feature/alert-action-reroute in [#133](https://github.com/iLert/terraform-provider-ilert/pull/133)